Renamed testshooter to wb_shooter_scout and added wb_shooter_infantry and wb_structure_prop. The infantry has a very similar role to the wmrts fighter melon, as it will likely be the "all-round" unit.
The prop structures is supposed to replace prop_physics apart from very specific exceptions,
as it is the same as a prop_physics with buildtime, health and belonging to a team.
Heavily inspired by the melon_baseprop from wmrts,
since its basically a necessary tweak for an rts to let players build their own tanks and such.
Otherwise everything but the "guns" will be invulnerable, and apart from being less fun, exploits would be "ahoy!!1".
Spent quite some time to make nice equations for the health of those props, with 3 separate equations,
one using mass, one using surface area and one using volume.
The health is then decided from the different results, so that the health is reasonable for most kinds of models.